solved the problems that were mentioned in the pool request

+++ src/main/java/core/basesyntax/IsoscelesTrapezoid.java
@@ -40,7 +40,7 @@ public class IsoscelesTrapezoid extends Figure {
 
     @Override
     public double getArea() {
-        return ((firstParallel + secondParallel) / 2) * height;
+        return ((double)(firstParallel + secondParallel) / 2) * height;
     }
 
     @Override
